pyproject.toml
[tool.poetry]
name = "idoven-ecg"
version = "0.1.0"
description = "ECG Analysis API"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
pydantic = {extras = ["email"], version = "^2.5.3"}
pydantic-settings = "^2.1.0"
alembic = "^1.13.0"
asyncpg = "^0.29.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["md5-crypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
celery = "^5.3.6"
httpx = "^0.26.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
pytest-mock = "^3.14.0"

[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
pytest-cov = "^4.1.0"
pytest = "^8.0.0"
pytest-asyncio = "^0.21.0"
faker = "^30.8.2"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
multi_line_output = 3
profile = "black"
line_length = 88
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true